@rafaelfranca 2021-01-05
This broke routes with symbols as the value of the constraints. Here is a test that fails:

```ruby
  def test_format_symbol_constraints
    rs.draw do
      get "/api", constraints: { format: :json },
                 to: lambda { |env| [200, {}, %w{json}] }
      get "/api", constraints: { format: :xml },
                 to: lambda { |env| [200, {}, %w{xml}] }
    end

    assert_equal "json", get(URI("http://www.example.org/api.json"))
    assert_equal "xml", get(URI("http://clients.example.org/api.xml"))
  end
```

I spend a few hours trying to figure out what exactly in this change broke that behavior but I couldn't. I'm reverting this change in 6ab985da28 adding the failing test so you can submit a new PR taking that in consideration. Sorry about that.