@theojulienne 2021-01-05
Thank you! I've fixed this in https://github.com/rails/rails/pull/41024